requirements.txt
requests==2.31.0
python-dotenv==1.0.1
certifi==2024.2.2
idna==3.6
urllib3==2.2.0